Update mapping_room.ino
Added a slash to finish a comment on line 45.
It looks great.
I think adding 'const'  to the array locks in the values; the size is locked by the initialization.

Notes from the reference page:
the address is an int; the value is a byte.
The EEPROM memory has a specified life of 100,000 write/erase cycles.
Locations that have never been written to have the value of 255.
int = EEPROM.read(address)
EEPROM.write(address, value) 
EEPROM.update(address, value) The value is written only if it is different from the one already saved at the same address.
EEPROM write takes 3.3 ms to complete. This will be important! example: For 64 rooms this means 211.2ms to write all data excluding internal processing time.
Get and Put can work with structures. If the usage data is in a structure, saving may be easier. Put uses update().

I think a separate initialization program should be written that sets the room counters to zero.
I think the room counters should use the word data type because I hope we have more than 255 uses. 
A word can store an unsigned number of at least 16 bits (from 0 to 65535) when displaying use HEX notation on the alphanumeric display 65535 = 0xFFFF.